consortium_crate::worker

Trait EventHandler

pub trait EventHandler: Send {
    // Provided methods
    fn on_start(&mut self, worker: &dyn Worker) { ... }
    fn on_read(&mut self, node: &str, fd: RawFd, msg: &[u8]) { ... }
    fn on_close(&mut self, node: &str, rc: i32) { ... }
    fn on_timeout(&mut self, node: &str) { ... }
    fn on_error(&mut self, node: &str, error: &WorkerError) { ... }
    fn take_buffers(
        &mut self,
    ) -> (HashMap<String, Vec<Vec<u8>>>, HashMap<String, Vec<Vec<u8>>>, HashSet<String>) { ... }
}

Event handler trait for worker events.

Implement this trait to receive notifications about worker events.

Provided Methods

fn on_start(&mut self, worker: &dyn Worker)

Called when the worker starts.

fn on_read(&mut self, node: &str, fd: RawFd, msg: &[u8])

Called when data is available for reading from a node.

fn on_close(&mut self, node: &str, rc: i32)

Called when a connection to a node is closed.

fn on_timeout(&mut self, node: &str)

Called when a node operation times out.

fn on_error(&mut self, node: &str, error: &WorkerError)

Called when an error occurs for a node.

fn take_buffers( &mut self, ) -> (HashMap<String, Vec<Vec<u8>>>, HashMap<String, Vec<Vec<u8>>>, HashSet<String>)

Extract buffered stdout/stderr data from this handler. Returns (stdout, stderr, timeouts) where stdout/stderr are node -> Vec<chunks>. Default returns empty maps. Used by Task to collect results after execution.